feat(back): refactor data exporting (MD format);
